fix: AI auto-tagging form fields
