+++ bookmarks/templates/settings/general.html
@@ -154,60 +154,40 @@ reddit.com/r/Music music reddit</pre>
               <span class="form-label d-inline-block">AI auto-tagging</span>
             </summary>
 
-            <div class="form-group{% if form.ai_api_key.errors %} has-error{% endif %}">
-              <label for="{{ form.ai_api_key.id_for_label }}" class="form-label">
-                API key
-              </label>
-              {{ form.ai_api_key|add_class:"form-input" }}
-              {% if form.ai_api_key.errors %}
-              <div class="form-input-hint is-error">
-                {{ form.ai_api_key.errors }}
-              </div>
-              {% endif %}
-              <div class="form-input-hint">
+            <div class="form-group">
+              {% formlabel form.ai_api_key "API key" %}
+              {% formfield form.ai_api_key has_help=True %}
+              {{ form.ai_api_key.errors }}
+              {% formhelp form.ai_api_key %}
                 Not required if using a compatible API without authentication.
-              </div>
-            </div>
-
-            <div class="form-group{% if form.ai_model.errors %} has-error{% endif %}">
-              <label for="{{ form.ai_model.id_for_label }}" class="form-label">
-                Model name
-              </label>
-              {{ form.ai_model|add_class:"form-input width-50 width-md-100" }}
-              {% if form.ai_model.errors %}
-              <div class="form-input-hint is-error">
-                {{ form.ai_model.errors }}
-              </div>
-              {% endif %}
-              <div class="form-input-hint">
-                <strong>Note:</strong> ensure the model you choose supports structured outputs.
-              </div>
-            </div>
-
-            <div class="form-group{% if form.ai_base_url.errors %} has-error{% endif %}">
-              <label for="{{ form.ai_base_url.id_for_label }}" class="form-label">
-                Base URL
-              </label>
-              {{ form.ai_base_url|add_class:"form-input width-50 width-md-100" }}
-              {% if form.ai_base_url.errors %}
-              <div class="form-input-hint is-error">
-                {{ form.ai_base_url.errors }}
-              </div>
-              {% endif %}
-              <div class="form-input-hint">
-                Leave blank to use OpenAI. You can use any OpenAI-compatible API, either self-hosted (Ollama: <code>http://localhost:11434/v1</code>) or cloud-based (Perplexity: <code>https://api.perplexity.ai</code>, OpenRouter: <code>https://openrouter.ai/api/v1</code>).
-              </div>
+              {% endformhelp %}
             </div>
 
             <div class="form-group">
-              <label for="{{ form.ai_tag_vocabulary.id_for_label }}" class="form-label">
-                Allowed tags
-              </label>
-              {{ form.ai_tag_vocabulary|add_class:"form-input monospace"|attr:"rows:8" }}
-              <div class="form-input-hint">
+              {% formlabel form.ai_model "Model name" %}
+              {% formfield form.ai_model has_help=True class="width-50 width-md-100" %}
+              {{ form.ai_model.errors }}
+              {% formhelp form.ai_model %}
+                <strong>Note:</strong> ensure the model you choose supports structured outputs.
+              {% endformhelp %}
+            </div>
+
+            <div class="form-group">
+              {% formlabel form.ai_base_url "Base URL" %}
+              {% formfield form.ai_base_url has_help=True class="width-50 width-md-100" %}
+              {{ form.ai_base_url.errors }}
+              {% formhelp form.ai_base_url %}
+                Leave blank to use OpenAI. You can use any OpenAI-compatible API, either self-hosted (Ollama: <code>http://localhost:11434/v1</code>) or cloud-based (Perplexity: <code>https://api.perplexity.ai</code>, OpenRouter: <code>https://openrouter.ai/api/v1</code>).
+              {% endformhelp %}
+            </div>
+
+            <div class="form-group">
+              {% formlabel form.ai_tag_vocabulary "Allowed tags" %}
+              {% formfield form.ai_tag_vocabulary has_help=True class="monospace" rows="8" %}
+              {% formhelp form.ai_tag_vocabulary %}
                 One tag per line. AI will <strong>only</strong> suggest tags from this list.<br>
                 <strong>Note:</strong> large tag lists increase API costs per request. The entire vocabulary is sent with each API call.
-              </div>
+              {% endformhelp %}
             </div>
           </details>
